requirements.txt
apiview==1.3.25
Django==4.2.17
django-cors-headers==4.6.0
django-import-export==4.3.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.10.1
serializers==0.2.4
sqlparse==0.5.3
drf-yasg==1.21.8
django-filter>=22.1